extract daemon gRPC client
