+++ src-tauri/core/src/connection/mod.rs
@@ -1,0 +1,1 @@
+pub mod daemon_client;
