+++ src-tauri/core/src/lib.rs
@@ -10,6 +10,7 @@ use std::{
 };
 
 pub mod app_config;
+pub mod connection;
 pub mod database;
 pub mod error;
 pub mod proxy;
